+++ src/three/rgb-image/annotated-box.ts
