+++ src/three/rgb-image/rgb-image.ts
